requirements.txt
Django==2.2.10
django-filter
djangorestframework
pytz==2021.3
sqlparse==0.4.2
drf-yasg2
